Change get/set music position to UINT32 parameter, milliseconds
# Conflicts:
#	src/d_netcmd.c
#	src/lua_baselib.c
#	src/nds/i_sound.c
#	src/sdl12/mixer_sound.c
#	src/sdl12/sdl_sound.c
#	src/win32ce/win_snd.c

+++ src/djgppdos/i_sound.c
@@ -550,13 +550,13 @@ boolean I_SetSongSpeed(float speed)
 	return false;
 }
 
-boolean I_SetSongPosition(float position)
+boolean I_SetSongPosition(UINT32 position)
 {
     (void)position;
     return false;
 }
 
-float I_GetSongPosition(void)
+UINT32 I_GetSongPosition(void)
 {
-    return 0.0f;
+    return 0.;
 }
