+++ src/dummy/i_sound.c
@@ -146,13 +146,13 @@ boolean I_SetSongTrack(int track)
 	return false;
 }
 
-boolean I_SetSongPosition(float position)
+boolean I_SetSongPosition(UINT32 position)
 {
     (void)position;
     return false;
 }
 
-float I_GetSongPosition(void)
+UINT32 I_GetSongPosition(void)
 {
-    return 0.0f;
+    return 0;
 }
